backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
PyMySQL==1.0.2
PyJWT==2.3.0
bcrypt==3.2.0
python-dotenv==0.19.0
Flask-Cors==3.0.10 